+++ src/main/drivers/serial_uart.h
@@ -66,7 +66,7 @@ serialPort_t *uartOpen(USART_TypeDef *USARTx, serialReceiveCallbackPtr callback,
 // serialPort API
 void uartWrite(serialPort_t *instance, uint8_t ch);
 uint32_t uartTotalRxBytesWaiting(serialPort_t *instance);
-uint8_t uartTotalTxBytesFree(serialPort_t *instance);
+uint32_t uartTotalTxBytesFree(serialPort_t *instance);
 uint8_t uartRead(serialPort_t *instance);
 void uartSetBaudRate(serialPort_t *s, uint32_t baudRate);
 bool isUartTransmitBufferEmpty(serialPort_t *s);
